feat(mcp): dedicated /mcp page + sidebar item; slim account to profile
- Add a "Developers → MCP" item to the sidebar and a /mcp route (under the app
  shell) with the Connect-to-Claude/Cursor instructions + API-key management,
  plus a Pro-gate notice for free users.
- Move that MCP/API-key UI out of the account page; account now shows profile
  (name, avatar, email) + plan/usage only.
